Add landfill json and handled clustering
For now the application is able to display a huge amount of cluster and icon on map. We got the cleaned wild dump, the waste, and the landfill. The next step is to create a separation of clusters between the three kinds of categories. In order to difference them and to manipulate them separately.